ingestion-service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-jose>=3.3.0
requests>=2.28.0
httpx>=0.24.0
pytest>=7.3.0
pypdf>=3.15.0
python-docx>=0.8.11
pytesseract>=0.3.10
Pillow>=10.0.0
langdetect>=1.0.9
numpy>=1.24.0
spacy>=3.6.0
boto3>=1.28.0
aiofiles>=0.8.0
python-magic>=0.4.27
redis>=4.5.0
celery>=5.3.0
flower>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.11.0 